More code cleanup, and tried to clean prim_fread some.

+++ proto2/src/look.c
@@ -840,7 +840,7 @@ do_examine(int descr, dbref player, const char *name, const char *dir)
     switch (Typeof(thing)) {
         case TYPE_ROOM:
             sprintf(buf, "%.*s" SYSNORMAL "  Owner: %s  Parent: ",
-                    (BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
+                    (int)(BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
                     ansi_unparse_object(OWNER(player), thing),
                     NAME(OWNER(thing)));
             strcat(buf,
@@ -849,20 +849,20 @@ do_examine(int descr, dbref player, const char *name, const char *dir)
             break;
         case TYPE_THING:
             sprintf(buf, "%.*s" SYSNORMAL "  Owner: %s  Value: %d",
-                    (BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
+                    (int)(BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
                     ansi_unparse_object(OWNER(player), thing),
                     NAME(OWNER(thing)), DBFETCH(thing)->sp.thing.value);
             break;
         case TYPE_PLAYER:
             sprintf(buf, "%.*s" SYSNORMAL "  %s: %d  ",
-                    (BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
+                    (int)(BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
                     ansi_unparse_object(OWNER(player), thing),
                     tp_cpennies, DBFETCH(thing)->sp.player.pennies);
             break;
         case TYPE_EXIT:
         case TYPE_PROGRAM:
             sprintf(buf, "%.*s" SYSNORMAL "  Owner: %s",
-                    (BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
+                    (int)(BUFFER_LEN - strlen(NAME(OWNER(thing))) - 35),
                     ansi_unparse_object(OWNER(player), thing),
                     NAME(OWNER(thing)));
             break;
